Refactor RouletteWheel select method to use classmethod
Refactor the `select` method in the `RouletteWheel` class to be a classmethod instead of an instance method. This change allows for easier usage of the method without the need to create an instance of the class. The `population` and `fitnesses` parameters are now passed as arguments to the method instead of being stored as instance variables. This improves encapsulation and makes the code more modular.